fix webhook removal and setup

+++ packages/contentful/scripts/remove-webhook.ts
@@ -1,20 +1,20 @@
 import * as contentful from 'contentful-management';
+import { getWebhook } from './setup-webhook';
 
 const spaceId = process.env.CONTENTFUL_SPACE_ID!;
 const contentfulManagementAccessToken =
   process.env.CONTENTFUL_MANAGEMENT_ACCESS_TOKEN!;
-const contentfulEnvironment = process.env.CONTENTFUL_ENVIRONMENT!;
 const client = contentful.createClient({
   accessToken: contentfulManagementAccessToken,
 });
 
 const app = async () => {
   const space = await client.getSpace(spaceId);
-  const webhook = await space.getWebhook(
-    `${contentfulEnvironment.toLowerCase()}-webhook`,
-  );
+  const webhook = await getWebhook(space);
 
-  await webhook.delete();
+  if (webhook) {
+    await webhook.delete();
+  }
 };
 
 app().catch((err) => {
